zcash_client_sqlite: Generalise `wallet::sapling` tests
These are all "single shielded pool" tests, and should pass for any
shielded pool implemented in the wallet.
